Removed group stories
