+++ packages/react/src/components/Card/Card.mdx
@@ -8,8 +8,6 @@ import {
 } from '@storybook/blocks';
 import { Information } from '../../../../../docs-components';
 import * as CardStories from './Card.stories';
-import * as CardGroupStories from './CardGroup/CardGroup.stories';
-import { Card } from '.';
 
 <Meta of={CardStories} />
 
@@ -53,7 +51,7 @@ subtil, og andre stiler.
 
 Demonstrerer hvordan flere Card komponenter kan grupperes sammen med `Card.Group`.
 
-<Canvas of={CardGroupStories.Preview} />
+<Canvas of={CardStories.Group} />
 
 <br />
