+++ packages/react/src/components/Card/Card.stories.tsx
@@ -1,9 +1,14 @@
 import React from 'react';
 import type { Meta, StoryFn } from '@storybook/react';
+import { TrashFillIcon } from '@navikt/aksel-icons';
 
 import media from '../../../../../assets/img/card-media.png';
-import { Heading, Paragraph } from '../Typography';
 import { Stack } from '../../../../../docs-components';
+import { Heading, Paragraph } from '../Typography';
+import { NativeSelect } from '../form/NativeSelect';
+import { Textfield } from '../form/Textfield';
+import { Button } from '../Button';
+import { Divider } from '../Divider';
 
 import { Card } from '.';
 
@@ -173,3 +178,81 @@ Variants.decorators = [
     </Stack>
   ),
 ];
+
+const options = [
+  { value: 'daglig leder', label: 'Dalig leder' },
+  { value: 'forretningsfører', label: 'Forretningsfører' },
+];
+
+export const Group: Story = (args) => (
+  <Card.Group {...args}>
+    <Card>
+      <Card.Header>
+        <Paragraph size='medium'>Rolle 1</Paragraph>
+        <Button
+          variant='secondary'
+          color='danger'
+          size='small'
+          icon={<TrashFillIcon aria-hidden />}
+        >
+          Fjern
+        </Button>
+      </Card.Header>
+      <Divider color='subtle' />
+      <Card.Content>
+        <NativeSelect label='Velg rolle'>
+          {options.map(({ value, label }, index) => (
+            <option
+              key={index}
+              value={value}
+            >
+              {label}
+            </option>
+          ))}
+        </NativeSelect>
+        <Textfield
+          size='small'
+          label='Fødsels- eller d-nummer'
+        />
+        <Textfield
+          size='small'
+          label='Etternavn'
+        />
+      </Card.Content>
+    </Card>
+    <Card>
+      <Card.Header>
+        <Paragraph size='medium'>Rolle 2</Paragraph>
+        <Button
+          variant='secondary'
+          color='danger'
+          size='small'
+          icon={<TrashFillIcon aria-hidden />}
+        >
+          Fjern
+        </Button>
+      </Card.Header>
+      <Divider color='subtle' />
+      <Card.Content>
+        <NativeSelect label='Velg rolle'>
+          {options.map(({ value, label }, index) => (
+            <option
+              key={index}
+              value={value}
+            >
+              {label}
+            </option>
+          ))}
+        </NativeSelect>
+        <Textfield
+          size='small'
+          label='Fødsels- eller d-nummer'
+        />
+        <Textfield
+          size='small'
+          label='Etternavn'
+        />
+      </Card.Content>
+    </Card>
+  </Card.Group>
+);
